--- src/resources/js/example.js
@@ -1,4 +1,0 @@
-function hello() {
-    console.log("yes it's hello");
-}
-export { hello };
